+++ R/tm_g_events_term_id.R
@@ -1,7 +1,6 @@
 #' Events by Term Plot Teal Module
 #'
 #' @description
-#' `r lifecycle::badge("stable")`
 #'
 #' Display Events by Term plot as a shiny module
 #'