+++ R/tm_g_heat_bygrade.R
@@ -1,7 +1,6 @@
 #' Teal module for the heatmap by grade
 #'
 #' @description
-#' `r lifecycle::badge("stable")`
 #'
 #' Display the heatmap by grade as a shiny module
 #'