+++ R/tm_g_spiderplot.R
@@ -1,7 +1,6 @@
 #' Spider plot Teal Module
 #'
 #' @description
-#' `r lifecycle::badge("stable")`
 #'
 #' Display spider plot as a shiny module
 #'